Fixed the bugs with the cutscene logic not setting the camera correctly.
Bug 1 - Added a buffer to the "trigger cutscene on scene start" logic so that the initial camera will get set correctly before starting the cutscene.

Bug 2 - When skipping a cutscene, the camera will not switch to whatever trigger the player last entered unless the cutscene is configured to do so.